requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
DateTime==4.3
discord.py==1.5.1
idna==3.0
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
pandas==1.2.0
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scikit-learn==0.24.0
scipy==1.6.0
six==1.15.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
zope.interface==5.2.0
